Add failing test reproducing datachannel send crash on client disconnect
Reproduces a crash seen on customer devices: when the client ends the call
while the device application is still sending datachannel messages, usrsctp
fails the send with ECONNRESET and libdatachannel throws
std::runtime_error("Sending failed, errno=..."). The exception leaks through
Datachannel::sendMessage(), so an application calling it without a try/catch
dies with std::terminate.

The new test establishes a real WebRTC connection against the device through
the virtual signaling stream, opens a datachannel, sends messages from an
application thread and lets the client hang up mid-stream. It asserts that
sendMessage() never throws and currently fails with the exception above.

The helper classes from signaling_tests.cpp are extracted to
test/test-common/virtual_device.hpp so both test files can share them.
